skip unauthorized replays
Updates the connector to skip replays that are not public, unless the uploader
`steam_id` matches that of the caller (in which case, we presumably will have access).
This avoids making API calls that result in a 404 response when we don't have access
to a replay. For some reason, these replays still get returned in a list response.